fix(audio): rename fullaudio bridge to FullAudioBridge
updated whitelist, also according to sonar checks
